Workaround bug #586122.
It seems that on some systems there is an assertion that gets tripped if you
try to use PyTypeObject.tp_traverse when the type is statically generated (not
on the heap). Seems to mostly affect Fedora, though any debug build is probably
affected.

This should work around it by using the inverse logic to skip traversal for
statically allocated Type objects.