one more fix needed for bug #586122.
The dump routine wasn't calling tp_traverse, but the get_referents one was.
I went a step further, though, and check whether the function is exactly the
one from PyType_Type. We may eventually need some sort of PyType_Check or
something like that, but this passes the basic tests for Max Kanat-Alexander.