Remove the collapse_instance_dicts safety check.
This means that we don't have to compute parents 2x.

I also determined (separately) why some dumps use so much memory for compute_parents,
it is because some objects can be referred to a huge number of times (check out None,
for example).
we should consider capping the parent lists (by default, but probably configurable).
It is pretty much impossible to actually make use of more than 100, and you have
to Proxy them to go via o.p anyway.