Fix test for whether an object has a GC head.
The scanner was testing whether an object has a GC head by testing its type
for the Py_TPFLAGS_HAVE_GC feature.  However, the proper test for whether an
object has a GC head also includes calling the type's tp_is_gc function if
it exists.  In most cases this makes no difference, but it matters for type
objects themselves, since some of those are statically-allocated and thus
cannot be garbage-collected.

Python provides a PyObject_IS_GC macro that implements the correct logic, so
use it.