Test scanner.size_of against sys.getsizeof.
The exact details of object sizes are difficult to maintain across Python
versions, and (as illustrated in the previous two commits) it's all too easy
for bugs to go undetected due to parallel implementation bugs in the code
under test and the test suite.

Python 2.6 introduced `sys.getsizeof`, so in all but a few special cases it
makes more sense to test against that.