Simplify Cython declarations using stock *.pxd files.
Cython provides *.pxd files with translated standard library declarations of
various kinds.  In general it makes more sense to use these rather than
declare equivalents ourselves, since it reduces the chance of bugs due to
errors in manual declarations, and it's easier to read.

This requires a few changes to non-import code in meliae/_loader.pyx due to
slight differences in whether some functions are declared to take "object"
or "PyObject *" arguments.  For this reason I've changed Py_INCREF and
Py_DECREF to Py_XINCREF and Py_XDECREF in several places, which is safe
because it only adds NULL checks; and a few places gain <object> casts,
which has the effect of adding Py_INCREF/Py_DECREF pairs but otherwise
doesn't affect the generated code.