Rework dump_object_info to write to file descriptors.
We want to minimise allocations here, and calling the write method is rather
too liable to make allocations.  For ordinary file objects (open() or
io.open()), writing directly to the file descriptor works in both Python 2
and 3.

The error handling isn't currently very good, but that was true with the
previous fwrite()-based code too (just better hidden).