Update _scanner.size_of for Python 3.
We now calculate the size of type objects using type.__sizeof__ (which
requires a little care, but works in both Python 2 and 3).

dict and unicode objects have been rearranged in various Python 3 versions
and the details are now substantially more complicated, so disable those
optimisations in the Python 3 case.

Finally, PyString_CheckExact can be spelled PyBytes_CheckExact as of Python
2.6, and the new spelling is required on Python 3.