Use more compact representations when loading dumps.
The natural representations of the "type_str", "name", and (in some cases)
"value" fields of objects loaded from dumps would be str, but on Python 3
this is a somewhat less efficient representation of ASCII strings, and for
dumps of large processes a compact representation may well matter.  To that
end, use bytes where possible.

In general I've tried to confine this to just the highest-volume objects.
For example, meliae.loader._TypeSummary doesn't need to be as dense, so
convenience makes more sense there and _TypeSummary.type_str is a str.

Some methods gain affordances to encode from or decode to str where
appropriate for convenience, where doing so doesn't cause other problems.

I extended meliae.tests.test_loader._example_dump to include both bytes and
text objects, in order to test the slightly different representations of
each.