Fix TestGetRecursiveSize.test_referenced for Python 3.
Python >= 3.3 uses a more compact representation of dicts whose keys are
Unicode strings; this tripped up this test.  Also, Python >= 3.7 does more
aggressive constant folding, defeating the attempt in this test to produce
two equal but non-object-identical strings.

Using unique objects as dict keys is sufficient for the purposes of this
test, and avoids both problems.